% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{checkExpression}
\alias{checkExpression}
\title{Check normalized expression
create scatter plot of normalized counts for a list of targeted genes}
\usage{
checkExpression(object, group, gridvar = NULL, target, ...)
}
\arguments{
\item{object}{An RNAqc instance}

\item{group}{grouping variable default to none}

\item{gridvar}{variable for facet grid, default to NULL}

\item{...}{settings for geom_point}

\item{genelist}{a vector of gene of interest}
}
\value{
scatter plot of the expression
}
\description{
Check normalized expression
create scatter plot of normalized counts for a list of targeted genes
}
