% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARsum.R
\name{Combine}
\alias{Combine}
\title{logCombine
It combines log columns of different samples}
\usage{
Combine(df1, df2)
}
\arguments{
\item{df1}{The star log file for sample 1}

\item{df2}{The star log file sample 2}
}
\value{
The joint df from df1 and df2
}
\description{
logCombine
It combines log columns of different samples
}
