% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{get_legend}
\alias{get_legend}
\title{get_legend
Helper function for getting legend}
\usage{
get_legend(gplot)
}
\arguments{
\item{gplot}{ggplot type of plot}
}
\value{

}
\description{
get_legend
Helper function for getting legend
}
