% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{getLegend}
\alias{getLegend}
\title{getLegend
Helper function for getting legend}
\usage{
getLegend(gplot)
}
\arguments{
\item{gplot}{ggplot type of plot}
}
\value{
legend element of plot
}
\description{
getLegend
Helper function for getting legend
}
