% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPIcounts.R
\name{getPIcounts}
\alias{getPIcounts}
\title{getPIcounts
Extract counting bases summary from PICARD output}
\usage{
getPIcounts(files, libnames)
}
\arguments{
\item{files}{all picard \code{.metrics} file}

\item{libnames}{Library Names}
}
\value{
joint table for counting bases from PICARD output
}
\description{
getPIcounts
Extract counting bases summary from PICARD output
}
