% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARcounts.R
\name{getSTARcounts}
\alias{getSTARcounts}
\title{getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table}
\usage{
getSTARcounts(rootdir, strand, level)
}
\arguments{
\item{rootdir}{The root directory under which the count file is stored}

\item{strand}{The strand-specific protocol used in RNA sequencing. Choose one from \code{c("first", "second", "unstranded")}}

\item{level}{indicate the structure of star output, level = 0 means basename contains libname, level = 1 and 2 indicates different nested dir structure.}
}
\value{
A strand-specific STAR count summary table
}
\description{
getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table
}
