% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARsum.R
\name{getSTARsum}
\alias{getSTARsum}
\title{getSTARsum
Produce a summary table for STAR alignment process}
\usage{
getSTARsum(files, libnames)
}
\arguments{
\item{files}{The list of files to read in}

\item{libnames}{Sample names, make sure they are paired with files}
}
\value{
A joint table for STAR alignment staitistics
}
\description{
getSTARsum
Produce a summary table for STAR alignment process
}
