% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPIcounts.R
\name{myPidfile}
\alias{myPidfile}
\title{myPidfile
get picard \code{.metrics} files from root directory}
\usage{
myPidfile(rootdir)
}
\arguments{
\item{rootdir}{Directory that stores all picard \code{.metrics} file}
}
\value{
A list contains all the \code{.metrics} files
}
\description{
myPidfile
get picard \code{.metrics} files from root directory
}
