% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARcounts.R
\name{mystarfile}
\alias{mystarfile}
\title{mystarfile
Generates the full path to a STAR count file.
It assume that the count file is of the form
rootdir/stardir/suffix}
\usage{
mystarfile(rootdir, stardir, suffix = "ReadsPerGene.out.tab")
}
\arguments{
\item{rootdir}{The root directory under which the STAR output is stored}

\item{stardir}{The directory holding the files generated by STAR}

\item{suffix}{Suffix used by STAR to identify countfile}
}
\value{
Full path to the STAR count file
}
\description{
mystarfile
Generates the full path to a STAR count file.
It assume that the count file is of the form
rootdir/stardir/suffix
}
