% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARsum.R
\name{mystarLogfile}
\alias{mystarLogfile}
\title{mystarLogfile
Generates the full path to a STAR log file.
It assume that the log file is of the form
rootdir/stardir/suffix}
\usage{
mystarLogfile(rootdir, stardir, suffix = "Log.final.out")
}
\arguments{
\item{rootdir}{The root directory under which the log file is stored}

\item{stardir}{The directory holding the files generated by STAR}

\item{suffix}{Suffix used by STAR to identify log file}
}
\value{
Full path to the STAR log file
}
\description{
mystarLogfile
Generates the full path to a STAR log file.
It assume that the log file is of the form
rootdir/stardir/suffix
}
