% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPIcounts.R
\name{picardCombine}
\alias{picardCombine}
\title{picardCombine
Row bind of picard summary for df1 and df2}
\usage{
picardCombine(df1, df2)
}
\arguments{
\item{df1}{picard summary for sample 1}

\item{df2}{picard summary for sample 2}
}
\value{
A joint summary dataframe from df1 and df2
}
\description{
picardCombine
Row bind of picard summary for df1 and df2
}
