% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\docType{methods}
\name{plotAlignment}
\alias{plotAlignment}
\alias{plot,RNAqc,missing-method}
\title{plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD}
\usage{
plotAlignment(object, group = NULL, textsize = 3, text = TRUE,
  angle = 60, ...)

\S4method{plot}{RNAqc,missing}(x, y, ...)
}
\arguments{
\item{object}{A RNAqc instance}

\item{group}{grouping variable, default to NULL. If specified, will also return scatter plot for mapping rate.}

\item{textsize}{Annotation size on the plot}

\item{angle}{x axis angle}

\item{...}{settings for geom_point}
}
\value{
A list of ggplot.
}
\description{
plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD
}
