% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{plotSTAR}
\alias{plotSTAR}
\alias{plotAll}
\alias{plot,RNAqc,missing-method}
\title{plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD}
\usage{
plotSTAR(object, textsize = 3, text = TRUE, angle = 60, ...)

plotAll(object, type, ...)

\S4method{plot}{RNAqc,missing}(x, y, ...)
}
\arguments{
\item{object}{an RNAqc instance}

\item{textsize}{Annotation size on the plot}

\item{angle}{x axis angle}

\item{...}{additional settings}

\item{type}{type of plot}

\item{group}{grouping variable, default to NULL. If specified, will also return scatter plot for mapping rate.}
}
\value{
A list of ggplot.
}
\description{
plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD
}
