% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Create Volcano plot for deseq2 results}
\usage{
plotVolcano(object, cutoff_nlog10p = -log10(0.05),
  cutoff_log2fc = log2(2), ...)
}
\arguments{
\item{object}{An RNAqc after model-fitting through DESeq2}

\item{cutoff_nlog10p}{Cutoff for adjusted values}

\item{cutoff_log2fc}{Cutoff for log fold change}

\item{...}{settings passed to results function from DESeq2}
}
\value{
a volcano plot
}
\description{
Create Volcano plot for deseq2 results
}
