% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAqc-class.R
\docType{methods}
\name{piData}
\alias{piData}
\alias{piData,RNAqc-method}
\alias{Nmap}
\alias{Nmap,RNAqc-method}
\alias{show,RNAqc-method}
\alias{piData<-}
\alias{piData<-,RNAqc-method}
\alias{Nmap<-}
\alias{Nmap<-,RNAqc-method}
\alias{[,RNAqc,ANY,ANY,ANY-method}
\alias{[<-,RNAqc,ANY,ANY,RNAqc-method}
\alias{coerce}
\alias{coerce,DESeqDataSet,RNAqc-method}
\title{Methods for RNAqc object}
\usage{
piData(object, ...)

\S4method{piData}{RNAqc}(object)

Nmap(object, ...)

\S4method{Nmap}{RNAqc}(object)

\S4method{show}{RNAqc}(object)

piData(object, ...) <- value

\S4method{piData}{RNAqc}(object) <- value

Nmap(object, ...) <- value

\S4method{Nmap}{RNAqc}(object) <- value

\S4method{[}{RNAqc,ANY,ANY,ANY}(x, i, j, drop = TRUE)

\S4method{[}{RNAqc,ANY,ANY,RNAqc}(x, i, j, ...) <- value
}
\arguments{
\item{object}{An RNAqc instance}
}
\description{
Methods for RNAqc object
}
