% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{starPCA}
\alias{starPCA}
\title{starPCA
Function for calculating PCA}
\usage{
starPCA(object, ntop = 1000)
}
\arguments{
\item{object}{an RNAqc object}

\item{ntop}{use top genes to calculate PCA}
}
\value{
dataframe ready for pca plot
}
\description{
starPCA
Function for calculating PCA
}
