% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{sumHTSeq}
\alias{sumHTSeq}
\title{sumHTSeq
Helper function that makes adjustments to output from getSTARcounts to be ready for plotting}
\usage{
sumHTSeq(counts)
}
\arguments{
\item{counts}{Counts table from STAR alignment}
}
\value{
A dataframe ready for plotting
}
\description{
sumHTSeq
Helper function that makes adjustments to output from getSTARcounts to be ready for plotting
}
