% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\docType{methods}
\name{summary.RNAqc}
\alias{summary.RNAqc}
\alias{summary,RNAqc-method}
\title{summary.RNAqc
summary of an RNAqc object to generate counts proportion table}
\usage{
summary.RNAqc(object)

\S4method{summary}{RNAqc}(object)
}
\arguments{
\item{object}{An RNAqc instance}
}
\value{
A dataframe ready for plotting
}
\description{
summary.RNAqc
summary of an RNAqc object to generate counts proportion table
}
