% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{createPCplot}
\alias{createPCplot}
\title{createPCplot
This function takes the meta data and star counts 
use DESeq2 for estimating size factors before considering any experiment designs.
It requires the column names of the input count data matrix and the row names of the metadata have the same order}
\usage{
createPCplot(object, var, labels = FALSE)
}
\arguments{
\item{object}{an RNAqc object}

\item{var}{grouping variables}

\item{labels}{show the labels or not}
}
\value{
PCA plot for pc1 vs pc2, pc1 vs pc3, pc2 vs pc3
}
\description{
createPCplot
This function takes the meta data and star counts 
use DESeq2 for estimating size factors before considering any experiment designs.
It requires the column names of the input count data matrix and the row names of the metadata have the same order
}
