% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARcounts.R
\name{getSTARcounts}
\alias{getSTARcounts}
\title{getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table}
\usage{
getSTARcounts(rootdir, strand)
}
\arguments{
\item{rootdir}{The root directory under which the count file is stored}

\item{strand}{The strand-specific protocol used in RNA sequencing. Choose one from \code{c("first", "second", "unstranded")}}
}
\value{
A strand-specific STAR count summary table
}
\description{
getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table
}
