% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARcounts.R
\name{getSTARcounts}
\alias{getSTARcounts}
\title{getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table
Make sure the files and library names are paired}
\usage{
getSTARcounts(files, libnames, strand, verbose = TRUE)
}
\arguments{
\item{files}{The list of files to read in}

\item{libnames}{Sample names, make sure they are paired with files}

\item{strand}{The strand-specific protocol used in RNA sequencing. Choose one from \code{c("first", "second", "unstranded")}}

\item{verbose}{True or False. Show read in progress}
}
\value{
A strand-specific STAR count summary table
}
\description{
getSTARcounts
Extract strand-specific STAR counts for each sample and then merge all counts into a single table
Make sure the files and library names are paired
}
