% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSTARsum.R
\name{getSTARsum}
\alias{getSTARsum}
\title{getSTARsum
Produce a summary table for STAR alignment process}
\usage{
getSTARsum(rootdir)
}
\arguments{
\item{rootdir}{The root directory under which the count file is stored}
}
\value{
A joint table for STAR alignment staitistics
}
\description{
getSTARsum
Produce a summary table for STAR alignment process
}
