% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{normalCounts}
\alias{normalCounts}
\title{Create normalized expression
generate dataset ready for scatter plot}
\usage{
normalCounts(object, target)
}
\arguments{
\item{object}{An RNAqc instance}

\item{target}{a vector of gene of interest}
}
\value{
dataframe for plotting
}
\description{
Create normalized expression
generate dataset ready for scatter plot
}
