% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{pcPlot}
\alias{pcPlot}
\title{pcPlot
Create PCA plot}
\usage{
pcPlot(df, vars, labels, pc)
}
\arguments{
\item{df}{input dataframe}

\item{vars}{grouping variables}

\item{labels}{sample id}

\item{pc}{principal components to compare}
}
\value{
ggplot type pca plot
}
\description{
pcPlot
Create PCA plot
}
