% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{PCplot}
\alias{PCplot}
\title{PCplot
Create PCA plot}
\usage{
PCplot(df, vars, labels, pc)
}
\arguments{
\item{df}{input dataframe}

\item{vars}{grouping variables}

\item{labels}{sample id}

\item{pc}{principal components to compare}
}
\value{
ggplot type pca plot
}
\description{
PCplot
Create PCA plot
}
