% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{plotAlignment}
\alias{plotAlignment}
\title{plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD}
\usage{
plotAlignment(object, group, textsize = 3, ...)
}
\arguments{
\item{object}{A RNAqc instance}

\item{textsize}{Annotation size on the plot}

\item{...}{Other visual options used for geom_text}

\item{type}{Counting reads or counting bases, select one from \code{c("HTSeq","Picard")}}
}
\value{
A ggplot type plot
}
\description{
plotAlignment
Plot stacked barplot for bases counting proportion for STAR and PICARD
}
