% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{plotSizefactor}
\alias{plotSizefactor}
\title{plotSizefactor
Create size factors generated by vst, equivalent as running estimateSizeFactors}
\usage{
plotSizefactor(object, group)
}
\arguments{
\item{object}{an RNAqc instance}

\item{group}{color grouping variable}
}
\value{
Sizefactor plot
}
\description{
plotSizefactor
Create size factors generated by vst, equivalent as running estimateSizeFactors
}
