% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAqc-class.R
\docType{class}
\name{RNAqc-class}
\alias{RNAqc-class}
\alias{.RNAqc}
\title{Create new RNAqc class
Inherit directly from SummarizedExperiment class}
\description{
Create new RNAqc class
Inherit directly from SummarizedExperiment class
}
