% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAqc-class.R
\docType{class}
\name{RNAqc-class}
\alias{RNAqc-class}
\alias{.RNAqc}
\alias{RNAqc-constructor}
\alias{RNAqc}
\title{Create new RNAqc class
Inherit directly from DESeqDataSet class}
\usage{
RNAqc(counts, colData, picard = DataFrame(), anno = NULL, ...)
}
\arguments{
\item{counts}{STAR counts matrix}

\item{colData}{manifest information}

\item{picard}{picard summary table}

\item{anno}{path to the annotation file}

\item{...}{for potential extra arguments in constructing SummarizedExperiment}
}
\description{
Create new RNAqc class
Inherit directly from DESeqDataSet class

Constructor for RNAqc Class
}
