% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAqc-class.R
\docType{class}
\name{RNAqc-class}
\alias{RNAqc-class}
\alias{.RNAqc}
\alias{RNAqc-constructor}
\alias{RNAqc}
\alias{normMatrix}
\alias{addGTF}
\title{Create new RNAqc class
Inherit directly from DESeqDataSet class}
\usage{
RNAqc(counts, colData, picard = DataFrame(), anno = NULL, ...)

normMatrix(obj)

addGTF(obj, gtfobj, id, type)
}
\arguments{
\item{counts}{STAR counts matrix}

\item{colData}{manifest information}

\item{picard}{picard summary table}

\item{anno}{path to the annotation file, first two columns must be gene id and gene symbols}

\item{...}{for potential extra arguments in constructing SummarizedExperiment}

\item{obj}{a RNAqc object}

\item{gtfobj}{a GRange object, typically import from rtracklayer}

\item{id}{gene ids or transcript ids}

\item{type}{regions from gtf to match}
}
\description{
Create new RNAqc class
Inherit directly from DESeqDataSet class

Constructor for RNAqc Class
}
