% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlignment.R
\name{summary.RNAqc}
\alias{summary.RNAqc}
\title{summary.RNAqc
Helper function that makes adjustments to output from getSTARcounts to be ready for plotting}
\usage{
\method{summary}{RNAqc}(object)
}
\arguments{
\item{object}{An RNAqc instance}
}
\value{
A dataframe ready for plotting
}
\description{
summary.RNAqc
Helper function that makes adjustments to output from getSTARcounts to be ready for plotting
}
